feature: make tsqmibadge its own component
