+++ src/pages/products/[product]/repositories/[repository]/Repository.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 
 import { NextPageWithLayout } from '@pages/_app.next';
 
-import { Box, Container } from '@mui/material';
+import { Box, Container, Tooltip } from '@mui/material';
 import SsidChartIcon from '@mui/icons-material/SsidChart';
 import LineAxisIcon from '@mui/icons-material/LineAxis';
 import SpeedIcon from '@mui/icons-material/Speed';
@@ -23,75 +23,18 @@ import CustomTabs from './components/CustomTabs';
 
 import { useQuery } from './hooks/useQuery';
 import OptionsHeader from './components/OptionsHeader/OptionsHeader';
-import { useRepositoryContext } from '@contexts/RepositoryProvider';
-import { toast } from 'react-toastify';
+import TsqmiBadge from './components/TsqmiBadge';
 
-const oneStarBadge = '/images/svg/badges/1stars.svg'
-const twoStarBadge = '/images/svg/badges/2stars.svg'
-const threeStarBadge = '/images/svg/badges/3stars.svg'
-const fourStarBadge = '/images/svg/badges/4stars.svg'
-const fiveStarBadge = '/images/svg/badges/5stars.svg'
-const zeroStarBadge = '/images/svg/badges/0stars.svg'
 
 const Repository: NextPageWithLayout = () => {
   useRequireAuth();
   useQuery();
 
-  const { latestTSQMI, latestTSQMIBadgeUrl } = useRepositoryContext();
-
   const [isHistoricCharacteristicOpen, setIsHistoricCharacteristicOpen] = useState(true);
   const [isHistoricSubCharacteristicOpen, setIsHistoricSubCharacteristicOpen] = useState(true);
   const [isHistoricMeasureOpen, setIsHistoricMeasureOpen] = useState(true);
 
-  const [showBadge, setShowBadge] = useState<boolean>(false);
-  const [badgePath, setBadgePath] = useState<any>('');
-
   const containerRef = React.useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    setStars();
-  }, [latestTSQMI])
-
-  const setStars = () => {
-    let star: any;
-
-    if (latestTSQMI) {
-      const value = latestTSQMI.value;
-
-      switch (true) {
-        case value == 0:
-          star = zeroStarBadge;
-          break;
-        case 0 < value && value < 0.2:
-          star = oneStarBadge;
-          break;
-        case 0.2 <= value && value < 0.4:
-          star = twoStarBadge;
-          break;
-        case 0.4 <= value && value < 0.6:
-          star = threeStarBadge;
-          break;
-        case 0.6 <= value && value < 0.8:
-          star = fourStarBadge;
-          break;
-        case 0.8 <= value && value <= 1.0:
-          star = fiveStarBadge;
-          break;
-        default:
-          star = null;
-      }
-    }
-
-    if (star) {
-      setBadgePath(star)
-      setShowBadge(true);
-    }
-  }
-
-  const copyToClipboard = () => {
-    navigator.clipboard.writeText(latestTSQMIBadgeUrl);
-    toast.success('Copiado com sucesso!');
-  }
 
   return (
     <Box display="flex" width="100%" flexDirection="row" marginTop="40px" marginBottom="24px">
@@ -99,18 +42,7 @@ const Repository: NextPageWithLayout = () => {
         <Box marginX="1%" maxWidth="98%">
           <Headers />
 
-          <Box
-            display="flex"
-            justifyContent="center"
-            sx={{
-              ':hover': {
-                cursor: 'pointer',
-              }
-            }}
-            onClick={copyToClipboard}
-          >
-            <img src={badgePath} alt="Exemplo SVG" style={{ width: '120px', height: '25px' }} />
-          </Box>
+          <TsqmiBadge />
 
           <OptionsHeader
             title='Características'
