+++ src/pages/products/[product]/repositories/[repository]/components/TsqmiBadge/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TsqmiBadge';
